- Fix credentials expire bug (#1) - Move session logic to aws/config.go
